refactor(ui): unify composer menu active-option scrolling
The slash-command and skill-mention composer menus each kept their own
byte-identical algorithm for scrolling the keyboard-active option into
view inside their ".slash-menu__scroll" viewport. Move that algorithm
into the shared chat-composer-dom module as
scrollActiveMenuOptionIntoView(activeId), and have each menu owner keep
resolving its own active-option id while delegating the scroll math to
the shared helper.

Fixes #124238
